refactor: extract tauri-free refresh_core into papr-core
Move the fetch/parse/ingest/newsletter/retention pipeline out of the
desktop scheduler into papr_core::ingestion::refresh::refresh_core, driven
by an on_event callback instead of a Tauri progress channel. The desktop
scheduler.rs is now a thin wrapper adding single-flight locking, the webview
progress channel, notifications, FreshRSS sync and tray updates.

RefreshSummary.ran preserves the 'idle Due tick stays idle' behaviour.
This lets the agent CLI run the exact same refresh headlessly.
