+++ crates/papr-core/Cargo.toml
@@ -14,6 +14,8 @@ rusqlite.workspace = true
 rusqlite_migration.workspace = true
 feed-rs.workspace = true
 reqwest.workspace = true
+tokio.workspace = true
+log.workspace = true
 scraper.workspace = true
 ego-tree.workspace = true
 url.workspace = true
