+++ crates/papr-core/src/ingestion/mod.rs
@@ -8,4 +8,5 @@ pub mod discovery;
 pub mod fetch;
 pub mod newsletter;
 pub mod parse;
+pub mod refresh;
 pub mod sources;
